Refactor config and enhance UI visuals
This commit encompasses a broad range of changes aimed at refining the application's configuration management and enhancing the user interface. Key modifications include the removal of debug-specific configuration loading for a more unified handling of sensitive information across build configurations. The addition of new bitmap resources in `Resources.Designer.cs` and updates to existing icons in `Resources.resx` significantly improve the visual elements of the application, catering to various resolutions and enhancing the overall user experience.

The project's assembly, file, and version numbers have been updated, reflecting the ongoing development and improvements. A shift in package references marks a notable change in the project's dependencies, notably the removal of `Microsoft.Extensions.Configuration.UserSecrets` in favor of `SixLabors.ImageSharp`, indicating a change in how images and possibly sensitive information are managed.

Significant updates to the UI include the addition of new menu items for "Mods" and "API Status", improving accessibility to these features. The introduction of a new configuration property `NexusApiKey` and related UI elements in the configuration form streamline the management of API keys. Enhancements in error handling and the addition of event handlers for API key authentication and management further refine the application's robustness and user interaction.

The removal of the `Secrets.cs` file and updates to how the Nexus API key is managed and authenticated against indicate a strategic move towards more secure and maintainable code practices. The introduction of a new static class `NexusApi` centralizes interactions with the Nexus API, simplifying the codebase and enhancing maintainability.

Overall, this commit represents a significant step forward in the application's development, focusing on improving configuration management, enhancing the UI, and refining the handling of sensitive information and API interactions.